Update Latin tokenizer to handle -cum compounds
Rewrote how the Latin tokenizer handles compounds of ablative
personal pronouns and the preposition 'cum'.

E.g. 'nobiscum' > ['cum', 'nobis']

The tokenizer now matches  words ending in '-cum' to a list of
valid "inclusions" (limited in number) rather than checking
against a list of "exceptions". The main  advantage is that
it no longer requires a list of exceptions to be maintained.