Updates to Latin word tokenizer (#311)
* Add feature to load Latin corpora directly using PlaintextCorpusReader

Using the Latin Library corpus as a test case, this feature allows
you to refer to a corpus and load it with PlaintextCorpusReader using
the following syntax:

from cltk.corpus.latin import latinlibrary

The code checks to make sure that a corpus is installed in the
main CLTK_DATA and raises an error if it is not there.

* Added missing comma to list 'cum_inclusions'

* Stop tokenizing 'nec'

This is based on the discussion here: https://github.com/PerseusDL/treebank_data/issues/8. Until this larger issue is resolved, it seems best to leave the form 'nec' as is.

* Added hyphen before tokenized enclitics

Looking at the Perseus NLP data, many of the tokenized enclitics
are distinguished using a hyphen, e.g. "-que". This brings it more
in line with that dataset. It also better distinguishes "ne" from "-ne".

* Add 'neque' to que_exceptions

Like 'nec', 'neque' should no longer be separated as two tokens to
bring the tokenizer more in line with Perseus NLP data (see
previous commit).

* Revert "Stop tokenizing 'nec'"

This reverts commit 1e80159024fdac209c8f527392fe4f96701203fe.

* Added hyphen before tokenized enclitics

Looking at the Perseus NLP data, many of the tokenized enclitics
are distinguished using a hyphen, e.g. "-que". This brings it more
in line with that dataset. It also better distinguishes "ne" from "-ne".

* Reversed order of tokenized enclitics

To better follow Perseus NLP practice, enclitics are now tokenized
in the order in which they appear, e.g.

'virumque' > ['virum', '-que']

See https://github.com/PerseusDL/treebank_data/issues/8

* Stop tokenizing 'nec'

This is based on the discussion here: https://github.com/PerseusDL/treebank_data/issues/8. Until this larger issue is resolved, it seems best to leave the form 'nec' as is.

* Better handle case for enclitic tokenization

* Updated test_tokenizer.py to reflect recent changes to the Latin tokenizer

* Rewrote "-cum" handling

Tokenization for "-cum" compounds, e.g. mecum, is now done through
regex replacement on the original string rather than by
iterating over and check all of the tokens. More efficient,
easier to read.

Includes a function to maintain case of original after
replacement.

* Handle exceptions in Latin tokenizer at string level

Following the logic of 'cum' compound handling, exceptional cases
are now handled through regex replacement at the string level.
This is much more efficient than the multiple list comprehensions
currently in use and much easier to read/maintain. It also now
correctly handles 'similist' and 'qualist'.

* Moved list of latin exceptions to a separate file

* Make tokenizer split final period; update test.

* Fixed typo in previous commit that will make Sanskrit tokenizer test fail

* Updated tokenizer to use local function instead of NLTK's word_tokenize

* Updated tokenizer to use local function instead of NLTK's word_tokenize

* Updates to Latin tokenizer

A few changes:
- Most significant: Special handling for Latin moved into its own
    function. Makes the general tokenizer code much easier to read
    and makes effort to avoid the clutter than will arise from
    separate exceptions for each language.
- Latin tokenizer now splits on sentences before splitting on words.
    This allows:
        - Better handling of '-ne' enclitic which can now be tested
            only on the sentence initial position.
        - Custom handling of Latin abbreviations. The test case
            included here are the praenomina; e.g. sentences will no
            longer incorrectly split on the name "Cn."