readable_tokens for build_concordance
Instead of returning a list of tokens as they occur in the main text, i added a case change for the string in question and "#_" for increasing visibility. As far as I know, the c-atf, used by Cuneiform Digital Library, and MdC-88, used by Thesaurus Linguae Aegyptiae, and Comprehensive Aramaic Lexicon, don't use "#_" in their encoding. I am not very good at it but it seems also regex compatible. I don't know its usage in greek-roman corpora though. In any case, i know that "#" or "_" by itself is used by the corpora I mentioned, ":","-" is also not good for the same reason, ":" represents a cuneiform character in elamite and old persian.