Added separator and file_name options
Here are the separator and file_name options we had discussed. 
The major problem was maintaining the flow of the word groups while changing file_name column. The solution I could come up with is mapping word groups to tokens and tokens to file names. The rest is, iterating over the word groups to produce the concordance line in relation to the filename associated with the token of the word group.