Clean up lookup.py
I used the `pylint` library to help streamline your classes. A few things
to point out: 1) If a class inherts from another, I believe it ought to call `super().__init__()` in order to reuse variables and functions. Not sure of the best way for you to do that here, however what I have done works. 2) I removed `load_replacement_patterns()` from `Synonyms`, because that in `Lemmata` is identical.