Transliterate Square Script to Imperial Aramaic
Simple transliteration function is added for transliterating
square script, modern hebrew, to imperial aramaic.
Going from imperial aramaic to modern hebrew is rather rare usecase
so it is not supported, besides the use of some characters
requires the knowledge of of its position within the word.
Some forms are word final etc.