Update grc.py
This is a solution to two bugs on the grc prosody function. The first changes the logic of the long by position so that the lack of either a mute OR a liquid allows for a long rather than the lack of them both. 
The second adds an enumerator to the scansion function so that locating the position of a syllable in a sentence is not reliant on the "index" function and therefore ought to handle reduplicated syllables better.